refactor: 이벤트 생성 모달 - 이벤트 알림 컴포넌트 분리
