+++ front/src/modal/component/CreateEventModal/AlertContainer/style.module.css
@@ -1,0 +1,3 @@
+.alert_button { cursor: pointer; width: fit-content; position: relative;  left: -10px; padding: 5px 10px; border-radius: 5px; } 
+.alert_button:hover { background: #f1f3f4; } 
+
